refactor(exception): custom exception 패키지 추가
