+++ src/test/java/gift/product/ProductControllerApiTest.java
@@ -5,11 +5,10 @@ import gift.member.presentation.dto.RequestMemberDto;
 import gift.product.persistence.entity.Category;
 import gift.product.persistence.entity.Product;
 import gift.product.persistence.repository.CategoryJpaRepository;
-import gift.product.persistence.repository.CategoryRepository;
 import gift.product.presentation.dto.RequestProductDto;
 import gift.product.presentation.dto.RequestProductIdsDto;
 import gift.product.presentation.dto.ResponseProductDto;
-import gift.global.exception.NotFoundException;
+import gift.global.exception.custrom.NotFoundException;
 import gift.product.persistence.repository.ProductRepository;
 import java.util.List;
 import java.util.Objects;
